added basic styling and some refatoring
