+++ src/Syringe.Web/Css/Syringe.scss
@@ -255,3 +255,12 @@
         }
     }
 }
+
+.drag-handle{
+    margin-right: 10px;
+    font: bold 20px Sans-Serif;
+    color: #5F9EDF;
+    display: inline-block;
+    cursor: move;
+    cursor: -webkit-grabbing;
+}
