+++ src/Syringe.Web/Scripts/Syringe/ReorderTest.ts
@@ -3,22 +3,6 @@
 module Syringe.Web {
     // declared Sortable so the sortable library can be used
     declare var Sortable: any;
-    export class TestPostion {
-
-        public OriginalPostion: number;
-        public Description: string;
-
-        constructor(originalPosition: number, description: string) {
-            this.OriginalPostion = originalPosition;
-            this.Description = description;
-        }
-    }
-
-    export class TestFilerOrder {
-        public Filename: string;
-        public Tests: TestPostion[];
-
-    }
 
     export class ReorderTests {
         constructor() {
@@ -47,12 +31,10 @@ module Syringe.Web {
 
                 var testFileOrder = new TestFilerOrder();
                 testFileOrder.Filename = $(this).data("filename");
-                var testPosition = new Array<TestPostion>();
-                $("#reorderedTestsList li").each(function(e) {
-                    testPosition.push(new TestPostion($(this).data("original"), ""));
-                });
 
-                testFileOrder.Tests = testPosition;
+                $("#reorderedTestsList li").each(function (e) {
+                    testFileOrder.Tests.push(new TestPostion($(this).data("original"), ""));
+                });
 
                 $.post("/TestFile/ReorderTests", { "testFile" : testFileOrder }, function(e) {
                     if (e) {
