+++ src/Syringe.Web/Scripts/Syringe/TestFilerOrder.ts
@@ -1,0 +1,10 @@
+module Syringe.Web {
+    export class TestFilerOrder {
+        public Filename: string;
+        public Tests: TestPostion[];
+
+        constructor() {
+            this.Tests = new Array<TestPostion>();
+        }
+    }
+}
