+++ src/Syringe.Web/Scripts/Syringe/TestPostion.ts
@@ -1,0 +1,12 @@
+module Syringe.Web {
+    export class TestPostion {
+
+        public OriginalPostion: number;
+        public Description: string;
+
+        constructor(originalPosition: number, description: string) {
+            this.OriginalPostion = originalPosition;
+            this.Description = description;
+        }
+    }
+}
