+++ src/Syringe.Web/Syringe.Web.csproj
@@ -432,6 +432,8 @@
     <Content Include="Views\TestFile\View.cshtml" />
     <TypeScriptCompile Include="Scripts\Syringe\ReorderTest.ts" />
     <TypeScriptCompile Include="Scripts\Syringe\RowHandler.ts" />
+    <TypeScriptCompile Include="Scripts\Syringe\TestFilerOrder.ts" />
+    <TypeScriptCompile Include="Scripts\Syringe\TestPostion.ts" />
     <TypeScriptCompile Include="Scripts\Syringe\ViewResult.ts" />
     <TypeScriptCompile Include="Scripts\typings\bootbox.d.ts" />
     <TypeScriptCompile Include="Scripts\typings\jquery\jquery.d.ts" />
